More changes for compute1

+++ scripts/hardware/30-kvm.sh
@@ -9,6 +9,12 @@ if [[ ! $INC_KVM ]]; then
     exit 0
 fi
 
+echo -e "\n--- Ensuring this isn't a VM ---\n"
+if [[ $(sudo dmidecode -s system-manufacturer) == 'QEMU' ]]; then
+    echo "This is a VM. Not installing KVM"
+    exit 0
+fi
+
 if [[ $OSNAME = "Fedora Linux" ]]; then
     echo "Installing KVM packages"
     sudo dnf install -y bridge-utils libvirt virt-install qemu-kvm libvirt-devel virt-top libguestfs-tools guestfs-tools virt-manager
@@ -19,11 +25,11 @@ fi
     sudo apt install -y qemu-kvm libvirt-daemon-system libvirt-clients bridge-utils ovmf
 fi
 
-echo "Stop and disable dnsmasq"
+echo -e "\n--- Stop and disable dnsmasq ---\n"
 sudo systemctl stop dnsmasq
 sudo systemctl disable dnsmasq
 
-echo "Creating libvirt group if it doesn't exist"
+echo -e "\n--- Creating libvirt group if it doesn't exist ---\n"
 sudo getent group | grep libvirt:
 
 if [[  $? -eq 1 ]]; then
@@ -31,18 +37,25 @@ if [[  $? -eq 1 ]]; then
     sudo groupadd --system libvirt
 fi
 
-echo "Adding ${USER} to libvirt group"
+echo -e "\n--- Adding ${USER} to libvirt group ---\n"
 sudo usermod -a -G libvirt ${USER}
 
-echo "Adding libvirt group to /etc/libvirt/libvirtd.conf"
+echo -e "\n--- Adding libvirt group to /etc/libvirt/libvirtd.conf ---\n"
 sudo sed -i 's/^#unix_sock_group = \"libvirt\"/unix_sock_group = \"libvirt\"/'
 sudo sed -i 's/^#unix_sock_rw_perms = \"0770\"/unix_sock_rw_perms = \"0770\"/'
+
+echo -e "\n--- Adding isos and vms Direcotry ---\n"
+sudo mkdir -p /var/lib/libvirt/images/isos
+sudo mkdir -p /var/lib/libvirt/images/vms
+
+#TODO Custom image directories
 
 # Set IOMMU if on
 if [[ ! -z $IOMMU ]]; then
    if [[ $CPU_VENDOR == 'AuthenticAMD' ]]; then
-       sudo sed -i -E 's/^(.*)"$/\1 amdiommu=on kvm.ignore_msrs=1"/' /etc/default/grub
+       sudo sed -i -E 's/^(GRUB_CMDLINE_LINUX_DEFAULT=".*)"$/\1 amdiommu=on kvm.ignore_msrs=1"/' /etc/default/grub
    fi
+
+   sudo grub-mkconfig -o /boot/grub/grub.cfg
 fi
 
-#TODO Custom image directories
